+++ src/core_modules/capture-core-utils/featuresSupport/support.js
@@ -4,7 +4,9 @@ export const FEATURES = Object.freeze({
     storeProgramStageWorkingList: 'storeProgramStageWorkingList',
     multiText: 'multiText',
     customIcons: 'customIcons',
+    newTransferQueryParam: 'newTransferQueryParam',
     exportablePayload: 'exportablePayload',
+    trackerImageEndpoint: 'trackerImageEndpoint',
 });
 
 // The first minor version that supports the feature
@@ -14,6 +16,8 @@ const MINOR_VERSION_SUPPORT = Object.freeze({
     [FEATURES.multiText]: 41,
     [FEATURES.customIcons]: 41,
     [FEATURES.exportablePayload]: 41,
+    [FEATURES.trackerImageEndpoint]: 41,
+    [FEATURES.newTransferQueryParam]: 41,
 });
 
 export const hasAPISupportForFeature = (minorVersion: string | number, featureName: string) =>