+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/EnrollmentAddEventPage.epics.js
@@ -10,6 +10,8 @@ import {
     commitEnrollmentEvents,
     rollbackEnrollmentEvents,
     saveFailed,
+    commitEnrollmentAndEvents,
+    rollbackEnrollmentAndEvents,
 } from '../common/EnrollmentOverviewDomain/enrollment.actions';
 import { actions as RelatedStageActions } from '../../WidgetRelatedStages/constants';
 import { buildUrlQueryString } from '../../../utils/routing';
@@ -103,5 +105,28 @@ export const saveNewEventFailedEpic = (action$: InputObservable) =>
         map((action) => {
             const { serverData } = action.meta;
             return batchActions([saveFailed(), rollbackEnrollmentEvents({ events: serverData.events })]);
+        }),
+    );
+
+export const saveEventAndCompleteEnrollmentSucceededEpic = (action$: InputObservable) =>
+    action$.pipe(
+        ofType(addEnrollmentEventPageDefaultActionTypes.EVENT_SAVE_ENROLLMENT_COMPLETE_SUCCESS),
+        map((action) => {
+            const meta = action.meta;
+            // the bundleReport returns the events in the same order as the payload order. Therefore, we know that the first event is the newly added one.
+            const eventId = action.payload.bundleReport.typeReportMap.EVENT.objectReports[0].uid;
+            return commitEnrollmentAndEvents(meta.uid, eventId);
+        }),
+    );
+
+export const saveEventAndCompleteEnrollmentFailedEpic = (action$: InputObservable) =>
+    action$.pipe(
+        ofType(addEnrollmentEventPageDefaultActionTypes.EVENT_SAVE_ENROLLMENT_COMPLETE_ERROR),
+        map((action) => {
+            const meta = action.meta;
+            return batchActions(
+                [saveFailed(), rollbackEnrollmentAndEvents(meta.uid)],
+                'NewEvent.saveEventAndCompleteEnrollmentFailed',
+            );
         }),
     );