+++ src/core_modules/capture-core/components/Pages/common/EnrollmentOverviewDomain/index.js
@@ -8,8 +8,13 @@ export {
     commitEnrollmentEvent,
     rollbackEnrollmentEvent,
     addEnrollmentEvents,
+    updateEnrollmentEventWithoutId,
     updateEnrollmentAttributeValues,
     showEnrollmentError,
+    updateEnrollmentAndEvents,
+    commitEnrollmentAndEvents,
+    rollbackEnrollmentAndEvents,
+    setExternalEnrollmentStatus,
 } from './enrollment.actions';
 export { useCommonEnrollmentDomainData } from './useCommonEnrollmentDomainData';
 export { useRuleEffects } from './useRuleEffects';