+++ src/core_modules/capture-core/components/WidgetEnrollment/TransferModal/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { TransferModal } from './TransferModal.component';