+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/Validated/validated.actions.js
@@ -19,6 +19,8 @@ export const newEventWidgetActionTypes = {
     START_CREATE_NEW_AFTER_COMPLETING: 'NewEvent.StartCreateNewAfterCompleting',
     SET_SAVE_ENROLLMENT_EVENT_IN_PROGRESS: 'NewEvent.SetSaveEnrollmentEventInProgress',
     CLEAN_UP_EVENT_SAVE_IN_PROGRESS: 'NewEvent.CleanUpDataEntry',
+    EVENT_SAVE_ENROLLMENT_COMPLETE_REQUEST: 'NewEvent.EventSaveAndEnrollmentCompleteRequest',
+    EVENT_SAVE_ENROLLMENT_COMPLETE: 'NewEvent.EventSaveAndEnrollmentComplete',
 };
 
 export const requestSaveEvent = ({
@@ -80,6 +82,85 @@ export const saveEvents = ({ serverData, onSaveErrorActionType, onSaveSuccessAct
 
 export const startCreateNewAfterCompleting = ({ enrollmentId, isCreateNew, orgUnitId, programId, teiId, availableProgramStages }: Object) =>
     actionCreator(newEventWidgetActionTypes.START_CREATE_NEW_AFTER_COMPLETING)({ enrollmentId, isCreateNew, orgUnitId, programId, teiId, availableProgramStages });
+
+
+export const requestSaveAndCompleteEnrollment = ({
+    eventId,
+    dataEntryId,
+    formFoundation,
+    programId,
+    orgUnitId,
+    orgUnitName,
+    teiId,
+    enrollmentId,
+    completed,
+    fromClientDate,
+    onSaveAndCompleteEnrollmentExternal,
+    onSaveAndCompleteEnrollmentSuccessActionType,
+    onSaveAndCompleteEnrollmentErrorActionType,
+    enrollment,
+}: {
+    eventId: string,
+    dataEntryId: string,
+    formFoundation: Object,
+    programId: string,
+    orgUnitId: string,
+    orgUnitName: string,
+    teiId: string,
+    enrollmentId: string,
+    completed?: boolean,
+    fromClientDate: (date: Date) => { getServerZonedISOString: () => string },
+    onSaveAndCompleteEnrollmentExternal?: (enrollmnet: ApiEnrollment) => void,
+    onSaveAndCompleteEnrollmentSuccessActionType?: string,
+    onSaveAndCompleteEnrollmentErrorActionType?: string,
+    enrollment: Object,
+}) =>
+    actionCreator(newEventWidgetActionTypes.EVENT_SAVE_ENROLLMENT_COMPLETE_REQUEST)(
+        {
+            eventId,
+            dataEntryId,
+            formFoundation,
+            programId,
+            orgUnitId,
+            orgUnitName,
+            teiId,
+            enrollmentId,
+            completed,
+            fromClientDate,
+            onSaveAndCompleteEnrollmentExternal,
+            onSaveAndCompleteEnrollmentSuccessActionType,
+            onSaveAndCompleteEnrollmentErrorActionType,
+            enrollment,
+        },
+        { skipLogging: ['formFoundation'] },
+    );
+
+export const saveEventAndCompleteEnrollment = (
+    serverData: Object,
+    onCompleteEnrollmentSuccessActionType?: string,
+    onCompleteEnrollmentErrorActionType?: string,
+    uid: string,
+) =>
+    actionCreator(newEventWidgetActionTypes.EVENT_SAVE_ENROLLMENT_COMPLETE)(
+        {},
+        {
+            offline: {
+                effect: {
+                    url: 'tracker?async=false&importStrategy=CREATE_AND_UPDATE',
+                    method: effectMethods.POST,
+                    data: serverData,
+                },
+                commit: onCompleteEnrollmentSuccessActionType && {
+                    type: onCompleteEnrollmentSuccessActionType,
+                    meta: { serverData, uid },
+                },
+                rollback: onCompleteEnrollmentErrorActionType && {
+                    type: onCompleteEnrollmentErrorActionType,
+                    meta: { serverData, uid },
+                },
+            },
+        },
+    );
 
 export const cleanUpEventSaveInProgress = () =>
     actionCreator(newEventWidgetActionTypes.CLEAN_UP_EVENT_SAVE_IN_PROGRESS)();