+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/Validated/validated.types.js
@@ -48,6 +48,7 @@ export type Props = {|
     itemId: string,
     onSave: (saveType: $Keys<addEventSaveTypes>) => void,
     onCancel: () => void,
+    onSaveAndCompleteEnrollment: (enrollment: ApiEnrollment) => void,
     formRef: (formInstance: any) => void,
     relatedStageRef: (relatedStageInstance: any) => void,
     dataEntryFieldRef: (instance: any, id: string) => void,