+++ src/core_modules/capture-core/reducers/descriptions/enrollmentDomain.reducerDescription.js
@@ -15,11 +15,14 @@ const {
     UPDATE_ENROLLMENT_EVENT,
     ADD_ENROLLMENT_EVENTS,
     UPDATE_ENROLLMENT_ATTRIBUTE_VALUES,
+    UPDATE_ENROLLMENT_AND_EVENTS,
     ROLLBACK_ENROLLMENT_EVENT,
     ROLLBACK_ENROLLMENT_EVENTS,
+    ROLLBACK_ENROLLMENT_AND_EVENTS,
     COMMIT_ENROLLMENT_EVENT,
     COMMIT_ENROLLMENT_EVENTS,
     ADD_PERSISTED_ENROLLMENT_EVENTS,
+    COMMIT_ENROLLMENT_AND_EVENTS,
 } = enrollmentSiteActionTypes;
 
 const setAssignee = (state, action) => {
@@ -131,6 +134,59 @@ export const enrollmentDomainDesc = createReducerDescription(
             ...state,
             attributeValues,
         }),
+        [UPDATE_ENROLLMENT_AND_EVENTS]: (state, { payload: { enrollment } }) => {
+            const enrollmentToUpdate = { ...enrollment, pendingApiResponse: true, dataToRollback: state.enrollment };
+
+            if (enrollmentToUpdate.events?.length > 0) {
+                const eventsToUpdate = state.enrollment.events.map((event) => {
+                    const eventToUpdate = enrollmentToUpdate.events.find(e => e.event === event.event);
+                    return eventToUpdate
+                        ? {
+                            ...eventToUpdate,
+                            pendingApiResponse: true,
+                            dataToRollback: event,
+                        }
+                        : event;
+                });
+                const eventWithoutId = enrollmentToUpdate.events.find(event => event.uid);
+                const eventsToUpdateAndAdd = eventWithoutId ?
+                    [...eventsToUpdate, { ...eventWithoutId, pendingApiResponse: true }] : eventsToUpdate;
+
+                return {
+                    ...state,
+                    enrollment: { ...state.enrollment, ...enrollmentToUpdate, events: eventsToUpdateAndAdd },
+                };
+            }
+
+            return { ...state, enrollment: { ...state.enrollment, ...enrollmentToUpdate } };
+        },
+        [ROLLBACK_ENROLLMENT_AND_EVENTS]: (state, { payload: { uid } }) => {
+            const enrollment = state.enrollment.dataToRollback;
+            const events = state.enrollment.events.reduce((acc, event) => {
+                if (uid && event.uid === uid) {
+                    return acc;
+                }
+                const eventToRollback = event.dataToRollback ? event.dataToRollback : event;
+                return [...acc, eventToRollback];
+            }, []);
+
+            return { ...state, enrollment: { ...state.enrollment, ...enrollment, events } };
+        },
+        [COMMIT_ENROLLMENT_AND_EVENTS]: (state, { payload: { eventId, uid } }) => {
+            const { pendingApiResponse, dataToRollback, ...enrollmentToCommit } = state.enrollment;
+            const eventsToCommit = state.enrollment.events.map((event) => {
+                const id = uid && event.uid === uid ? eventId : event.event;
+                const {
+                    pendingApiResponse: pendingApiResponseEvent,
+                    dataToRollback: dataToRollbackEvent,
+                    uid: uidToRemove,
+                    ...dataToCommit
+                } = event;
+                return { ...dataToCommit, event: id };
+            });
+
+            return { ...state, enrollment: { ...enrollmentToCommit, events: eventsToCommit } };
+        },
         [enrollmentNoteActionTypes.ADD_ENROLLMENT_NOTE]:
         (state, { payload: { note } }) => ({
             ...state,